Fix import route for getActivePage
